Split to classes
