+++ app/Filament/Blocks/RichContentBlocks/ModulesSection.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Filament\Blocks\RichContentBlocks;
+use Filament\Forms;
+use Filament\Forms\Form;
+
+class ModulesSection extends BaseBlock
+{
+    static function schema(Form $form)
+    {
+        return [
+            Forms\Components\TextInput::make('per_page')->numeric()->default(9),
+        ];
+    }
+}
